add loggers.py in registry, remove code from parser.py
